fix(security): tambah batas max goods_value_idr + int constraint pada COD fee percent
QA security review menemukan 2 MEDIUM:
1. goods_value_idr tanpa upper bound (3 skema) → tambah max(1_000_000_000)
2. JNE/SAP_COD_FEE_PERCENT tidak dikonstrain integer → tambah .int()
3. Lint: unused var 'service' di destination-resolution test → ganti void

+++ src/config/env.ts
@@ -33,7 +33,7 @@ export const envSchema = z.object({
   JNE_SHIPPER_ZIP: optionalSecret,
   JNE_WEBHOOK_TOKEN: optionalSecret,
   JNE_COD_ELIGIBLE_SERVICES: z.string().default('REG,JTR'),
-  JNE_COD_FEE_PERCENT: z.coerce.number().min(0).max(100).default(3),
+  JNE_COD_FEE_PERCENT: z.coerce.number().int().min(0).max(100).default(3),
   JNE_COD_MIN_FEE_IDR: z.coerce.number().int().nonnegative().default(0),
   SAP_API_BASE_URL: optionalUrl,
   SAP_TRACKING_BASE_URL: optionalUrl,
@@ -50,7 +50,7 @@ export const envSchema = z.object({
   SAP_SHIPPER_PHONE: optionalSecret,
   SAP_SHIPPER_CONTACT: optionalSecret,
   SAP_WEBHOOK_TOKEN: optionalSecret,
-  SAP_COD_FEE_PERCENT: z.coerce.number().min(0).max(100).default(2),
+  SAP_COD_FEE_PERCENT: z.coerce.number().int().min(0).max(100).default(2),
   SAP_COD_MIN_FEE_IDR: z.coerce.number().int().nonnegative().default(0),
 })
